fastapi/requirements.txt
torch==2.3.1+cu118
torchvision==0.18.1+cu118
torchaudio==2.3.1+cu118
-f https://download.pytorch.org/whl/torch_stable.html
fastapi==0.112.0
uvicorn==0.30.5
databases
aiomysql
motor
itsdangerous==2.2.0
python-dotenv==1.0.1
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
setuptools==65.5.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
requests==2.32.3
httpx==0.27.0
pytest
pytest-asyncio
langchain-community
transformers
bitsandbytes
accelerate>=0.26.0
ua-parser
jpype1
konlpy
googletrans
langdetect
beautifulsoup4
networkx
databases